refactor(orc8r): Segregation of state.proto & its servicers
