+++ cwf/cloud/go/services/cwf/obsidian/handlers/handlers_test.go
@@ -876,7 +876,7 @@ func seedCwfNetworks(t *testing.T) {
 }
 
 func reportSubscriberDirectoryRecord(t *testing.T, ctx context.Context, id string, req *directorydTypes.DirectoryRecord) {
-	client, err := state.GetStateClient()
+	client, err := state.GetExternalStateClient()
 	assert.NoError(t, err)
 
 	serializedRecord, err := serde.Serialize(req, orc8r.DirectoryRecordType, serdes.State)
@@ -897,7 +897,7 @@ func reportSubscriberDirectoryRecord(t *testing.T, ctx context.Context, id strin
 }
 
 func reportHaPairStatus(t *testing.T, ctx context.Context, pairID string, req *models2.CarrierWifiHaPairStatus) {
-	client, err := state.GetStateClient()
+	client, err := state.GetExternalStateClient()
 	assert.NoError(t, err)
 
 	serializedRecord, err := serde.Serialize(req, cwf.CwfHAPairStatusType, serdes.State)
@@ -918,7 +918,7 @@ func reportHaPairStatus(t *testing.T, ctx context.Context, pairID string, req *m
 }
 
 func reportGatewayHealthStatus(t *testing.T, ctx context.Context, gatewayID string, req *models2.CarrierWifiGatewayHealthStatus) {
-	client, err := state.GetStateClient()
+	client, err := state.GetExternalStateClient()
 	assert.NoError(t, err)
 
 	serializedRecord, err := serde.Serialize(req, cwf.CwfGatewayHealthType, serdes.State)
