+++ orc8r/cloud/go/services/state/protos/state.proto
@@ -1,0 +1,57 @@
+/*
+Copyright 2020 The Magma Authors.
+
+This source code is licensed under the BSD-style license found in the
+LICENSE file in the root directory of this source tree.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+syntax = "proto3";
+package magma.orc8r.state;
+
+import "orc8r/protos/service303.proto";
+import "orc8r/protos/state.proto";
+
+option go_package = "magma/orc8r/cloud/go/services/state/protos";
+
+// GetStatesRequest functions in two modes
+//  - ids is non-empty  -- normal Get
+//  - ids empty         -- Search with filters and load criteria
+message GetStatesRequest {
+    // networkID of the network containing desired state.
+    string networkID = 1;
+
+    // ids for which to return corresponding state.
+    repeated magma.orc8r.StateID ids = 2;
+
+    // typeFilter is an OR of all types specified.
+    // The service will return all states matching any of the specified types.
+    repeated string typeFilter = 10;
+
+    // idFilter is an OR of all device IDs specified.
+    // The service will return all states matching any of the specified IDs.
+    repeated string idFilter = 11;
+
+    // loadValues specifies whether state values should be loaded in
+    // addition to metadata.
+    bool loadValues = 12;
+
+    // If non-empty, the value of id_prefix will be used to do a prefix-match
+    // on the keys of queried states. This argument supersedes any value for
+    // idFilter.
+    string id_prefix = 13;
+}
+
+message GetStatesResponse {
+    repeated magma.orc8r.State states = 1;
+}
+
+service StateInternalService {
+    // GetStates retrieves states from blobstorage.
+    rpc GetStates (GetStatesRequest) returns (GetStatesResponse) {}
+}
