+++ orc8r/protos/state.proto
@@ -24,38 +24,6 @@ message StateID {
     string deviceID = 2;
 }
 
-// GetStatesRequest functions in two modes
-//  - ids is non-empty  -- normal Get
-//  - ids empty         -- Search with filters and load criteria
-message GetStatesRequest {
-    // networkID of the network containing desired state.
-    string networkID = 1;
-
-    // ids for which to return corresponding state.
-    repeated StateID ids = 2;
-
-    // typeFilter is an OR of all types specified.
-    // The service will return all states matching any of the specified types.
-    repeated string typeFilter = 10;
-
-    // idFilter is an OR of all device IDs specified.
-    // The service will return all states matching any of the specified IDs.
-    repeated string idFilter = 11;
-
-    // loadValues specifies whether state values should be loaded in
-    // addition to metadata.
-    bool loadValues = 12;
-
-    // If non-empty, the value of id_prefix will be used to do a prefix-match
-    // on the keys of queried states. This argument supersedes any value for
-    // idFilter.
-    string id_prefix = 13;
-}
-
-message GetStatesResponse {
-    repeated State states = 1;
-}
-
 message ReportStatesRequest {
     repeated State states = 1;
 }
@@ -89,9 +57,6 @@ message SyncStatesResponse {
 }
 
 service StateService {
-    // GetStates retrieves states from blobstorage.
-    rpc GetStates (GetStatesRequest) returns (GetStatesResponse) {}
-
     // ReportStates saves states into blobstorage.
     rpc ReportStates(ReportStatesRequest) returns (ReportStatesResponse) {}
